refactor: 🚨 fixed dependency loop + fixed SVCs with gaps during the day to show "resume" status properly
